Convert functions to use int64

+++ is.go
@@ -1,7 +1,7 @@
 package prime
 
 // Is determines whether a given number is prime.
-func Is(n int) bool {
+func Is(n int64) bool {
 	if n <= 1 {
 		return false
 	}
@@ -14,7 +14,7 @@ func Is(n int) bool {
 		return false
 	}
 
-	for i := 5; i*i <= n; i += 6 {
+	for i := int64(5); i*i <= n; i += 6 {
 		if n%i == 0 || n%(i+2) == 0 {
 			return false
 		}
